requirements.txt
fastapi
uvicorn
peewee
passlib[bcrypt]
python-dotenv
apscheduler
requests
google-genai
groq
openai
transformers
torch
sentencepiece
python-multipart
PyJWT
fastapi
uvicorn
python-dotenv
openai
pydantic
sqlalchemy